Module trust_dns_client::op

Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used together to either query or update resource records sets.

Re-exports

pub use self::update_message::UpdateMessage;

Modules

update_message

Update related operations for Messages

Structs

DnsResponse

A DNS response object

Edns

Edns implements the higher level concepts for working with extended dns as it is used to create or be created from OPT record data.

Header

Metadata for the Message struct.

LowerQuery

Identical to [trust_dns_client::op::Query], except that the Name is guaranteed to be in lower case form

Message

The basic request and response datastructure, used for all DNS protocols.

Query

Query struct for looking up resource records, basically a resource record without RDATA.

Enums

MessageType

Message types are either Query (also Update) or Response

OpCode

Operation code for queries, updates, and responses

ResponseCode

The status code of the response to a query.

Traits

MessageFinalizer

A trait for performing final amendments to a Message before it is sent.