Struct trust_dns_client::rr::dnssec::SupportedAlgorithms

pub struct SupportedAlgorithms { /* private fields */ }

Used to specify the set of SupportedAlgorithms between a client and server

Implementations

impl SupportedAlgorithms

pub fn new() -> SupportedAlgorithms

Return a new set of Supported algorithms

pub fn all() -> SupportedAlgorithms

Specify the entire set is supported

pub fn from_vec(algorithms: &[Algorithm]) -> SupportedAlgorithms

Based on the set of Algorithms, return the supported set

pub fn set(&mut self, algorithm: Algorithm)

Set the specified algorithm as supported

pub fn has(self, algorithm: Algorithm) -> bool

Returns true if the algorithm is supported

pub fn iter(&self) -> SupportedAlgorithmsIter<'_>

Return an Iterator over the supported set.

pub fn len(self) -> u16

Return the count of supported algorithms

pub fn is_empty(self) -> bool

Return true if no SupportedAlgorithms are set, this implies the option is not supported