Function trust_dns_client::rr::dnssec::tbs::rrset_tbs_with_sig

pub fn rrset_tbs_with_sig(
    name: &Name,
    dns_class: DNSClass,
    sig: &SIG,
    records: &[Record]
) -> Result<TBS, ProtoError>

Returns the to-be-signed serialization of the given record set using the information provided from the SIG record.

Arguments

name - labels of the record to sign

dns_class - DNSClass of the RRSet, i.e. IN

sig - SIG or RRSIG record, which was produced from the RRSet

records - RRSet records to sign with the information in the rrsig

Return

binary hash of the RRSet with the information from the RRSIG record