Module trust_dns_client::rr::rdata

All record data structures and related serialization methods

Modules

a

IPv4 address record data

aaaa

IPv6 address record data

caa

allows a DNS domain name holder to specify one or more Certification Authorities (CAs) authorized to issue certificates for that domain.

dnskey

public key record data for signing zone records

ds

pointer record from parent zone to child zone for dnskey proof

hinfo

HINFO record for storing host information

key

public key record data for signing zone records

mx

mail exchange, email, record

name

Record type for all cname like records.

naptr

Dynamic Delegation Discovery System

nsec

negative cache proof for non-existence

nsec3

hashed negative cache proof for non-existence

nsec3param

parameters used for the nsec3 hash method

null

null record type, generally not used except as an internal tool for representing null data

openpgpkey

OPENPGPKEY records for OpenPGP public keys

opt

option record for passing protocol options between the client and server

sig

signature record for signing queries, updates, and responses

soa

start of authority record defining ownership and defaults for the zone

srv

service records for identify port mapping for specific services on a host

sshfp

SSHFP records for SSH public key fingerprints

svcb

SVCB records, see draft-ietf-dnsop-svcb-https-03 SVCB and HTTPS RRs for DNS, February 2021

tlsa

TLSA records for storing TLS certificate validation information

txt

text records for storing arbitrary data

Structs

CAA

The CAA RR Type

DNSKEY

RFC 4034, DNSSEC Resource Records, March 2005

DS

RFC 4034, DNSSEC Resource Records, March 2005

HINFO

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

KEY

RFC 2535, Domain Name System Security Extensions, March 1999

MX

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

NAPTR

RFC 3403 DDDS DNS Database, October 2002

NSEC

RFC 4034, DNSSEC Resource Records, March 2005

NSEC3

RFC 5155, NSEC3, March 2008

NSEC3PARAM

RFC 5155, NSEC3, March 2008

NULL

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

OPENPGPKEY

RFC 7929

OPT

The OPT record type is used for ExtendedDNS records.

SIG

RFC 2535, Domain Name System Security Extensions, March 1999

SOA

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

SRV

RFC 2782, DNS SRV RR, February 2000

SSHFP

RFC 4255

SVCB

draft-ietf-dnsop-svcb-https-03 SVCB and HTTPS RRs for DNS, February 2021

TLSA

RFC 6698, DNS-Based Authentication for TLS

TXT

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

Enums

DNSSECRData

Record data enum variants for DNSSEC-specific records.

DNSSECRecordType

The type of the resource record, for DNSSEC-specific records.